+++ config/spam_detector/akismet.php
@@ -1,8 +1,8 @@
 <?php
+
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
 use FOS\ChatBundle\Service\SpamDetection\AkismetSpamDetector;
-
 
 return function(ContainerConfigurator $container): void 
 {
@@ -10,10 +10,13 @@ return function(ContainerConfigurator $container): void
 
         ->set('fos_chat.akismet_spam_detector', AkismetSpamDetector::class)
             ->args([                        
-                // service(''), //TODO: implement
+                env('AKISMET_API_KEY'),
+                env('AKISMET_SITE_URL')->default(''),
+                service('http_client'),
+                service('request_stack'),
                 service('fos_chat.participant_provider'),
             ])
             ->alias(AkismetSpamDetector::class, 'fos_chat.akismet_spam_detector')
             ->alias('fos_chat.spam_detector', 'fos_chat.akismet_spam_detector')
     ;
-}; 
+};