+++ src/Service/SpamDetection/SpamDetectorInterface.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace FOS\ChatBundle\Service\SpamDetection;
 
+use FOS\ChatBundle\Model\SpamStatus;
+
 /**
  * Tells whether or not a new message looks like spam.
  *
@@ -11,8 +13,5 @@ namespace FOS\ChatBundle\Service\SpamDetection;
  */
 interface SpamDetectorInterface
 {
-    /**
-     * Tells whether or not a new message looks like spam.
-     */
-    public function isSpam($message) : bool;
+    public function check(string $message): SpamStatus;
 }